requirements.txt
discord.py==0.16.12
aiohttp
websockets
chardet
asyncio